requirements.txt
aws-cdk-lib==2.99.1
constructs>=10.0.0,<11.0.0
boto3
langchain
pypdf
chromadb
streamlit
anthropic